rename folder + file and refactor code
