--- jobfair-frontend/src/fake-data/employees.json
@@ -1,239 +1,0 @@
-[
-    {
-      "accountId": "79dc6dea-8a7d-4ecc-a5cf-6eeb843467f0",
-      "account": {
-        "id": "79dc6dea-8a7d-4ecc-a5cf-6eeb843467f0",
-        "email": "manager@apple.com",
-        "password": "$2a$10$ni.o7mcmOw8SW5CF5WpRaeWwFmU9xs2cdxgzjxepXQXjWYDjX6yvO",
-        "status": "VERIFIED",
-        "phone": "0917499917",
-        "profileImageUrl": "https://d3polnwtp0nqe6.cloudfront.net/default.png",
-        "firstname": "(Apple) Bui ",
-        "lastname": "Khoi (Manager)",
-        "middlename": "Minh",
-        "gender": "MALE",
-        "role": "COMPANY_MANAGER"
-      },
-      "companyDTO": {
-        "id": "8e407290-9bdb-4e12-b7d3-d1ffdd1d8479",
-        "taxId": "18002752273",
-        "name": "Apple",
-        "address": "Apple",
-        "phone": "0917499917",
-        "email": "email@apple.com",
-        "employeeMaxNum": 5,
-        "websiteUrl": "https://www.apple.com",
-        "status": "REGISTERED",
-        "sizeId": 0,
-        "subCategoryDTOs": [
-          {
-            "id": 57,
-            "name": "NGO/Non-profit",
-            "category": {
-              "id": 12,
-              "name": "Services"
-            }
-          },
-          {
-            "id": 10,
-            "name": "Overseas jobs",
-            "category": {
-              "id": 1,
-              "name": "Type"
-            }
-          },
-          {
-            "id": 14,
-            "name": "Sales",
-            "category": {
-              "id": 2,
-              "name": "Front Office"
-            }
-          }
-        ],
-        "companyBenefitDTOS": [
-          {
-            "id": 51,
-            "description": "$200 a month kindergarten allowance per child under 5 years old",
-            "benefitDTO": {
-              "id": 0,
-              "name": "Kindergarten",
-              "description": "e.g: $200 a month kindergarten allowance per child under 5 years old"
-            },
-            "companyDTO": null
-          },
-          {
-            "id": 52,
-            "description": "Lunch and unlimited snacks",
-            "benefitDTO": {
-              "id": 1,
-              "name": "Canteen",
-              "description": "e.g: Lunch and unlimited snacks"
-            },
-            "companyDTO": null
-          }
-        ],
-        "mediaDTOS": [],
-        "companyLogoURL": "https://d3polnwtp0nqe6.cloudfront.net/default.png",
-        "companyDescription": "Apple đại đế"
-      }
-    },
-    {
-      "accountId": "c4fba70f-09d3-400b-b062-c00f96cec9e4",
-      "account": {
-        "id": "c4fba70f-09d3-400b-b062-c00f96cec9e4",
-        "email": "employee@apple.com",
-        "password": "$2a$10$pz.P0cOnN58fu3VQI683Z.TcV9Ys6pm592H0.OKzszdVBTUrJPULq",
-        "status": "VERIFIED",
-        "phone": "0917499917",
-        "profileImageUrl": "https://d3polnwtp0nqe6.cloudfront.net/default.png",
-        "firstname": "(Apple) Bui",
-        "lastname": "Khoi (Epmloyee)",
-        "middlename": "Minh",
-        "gender": "MALE",
-        "role": "COMPANY_EMPLOYEE"
-      },
-      "companyDTO": {
-        "id": "8e407290-9bdb-4e12-b7d3-d1ffdd1d8479",
-        "taxId": "18002752273",
-        "name": "Apple",
-        "address": "Apple",
-        "phone": "0917499917",
-        "email": "email@apple.com",
-        "employeeMaxNum": 5,
-        "websiteUrl": "https://www.apple.com",
-        "status": "REGISTERED",
-        "sizeId": 0,
-        "subCategoryDTOs": [
-          {
-            "id": 57,
-            "name": "NGO/Non-profit",
-            "category": {
-              "id": 12,
-              "name": "Services"
-            }
-          },
-          {
-            "id": 10,
-            "name": "Overseas jobs",
-            "category": {
-              "id": 1,
-              "name": "Type"
-            }
-          },
-          {
-            "id": 14,
-            "name": "Sales",
-            "category": {
-              "id": 2,
-              "name": "Front Office"
-            }
-          }
-        ],
-        "companyBenefitDTOS": [
-          {
-            "id": 51,
-            "description": "$200 a month kindergarten allowance per child under 5 years old",
-            "benefitDTO": {
-              "id": 0,
-              "name": "Kindergarten",
-              "description": "e.g: $200 a month kindergarten allowance per child under 5 years old"
-            },
-            "companyDTO": null
-          },
-          {
-            "id": 52,
-            "description": "Lunch and unlimited snacks",
-            "benefitDTO": {
-              "id": 1,
-              "name": "Canteen",
-              "description": "e.g: Lunch and unlimited snacks"
-            },
-            "companyDTO": null
-          }
-        ],
-        "mediaDTOS": [],
-        "companyLogoURL": "https://d3polnwtp0nqe6.cloudfront.net/default.png",
-        "companyDescription": "Apple đại đế"
-      }
-    },
-    {
-      "accountId": "f996f062-c1d5-4fcc-a13c-8cae1e210d0d",
-      "account": {
-        "id": "f996f062-c1d5-4fcc-a13c-8cae1e210d0d",
-        "email": "employee2@apple.com",
-        "password": "$2a$10$Lcnrap7QluPxEiKc6OQEOepWi6IGiV4mAlqNQ3qcGw83R09uo.9Aa",
-        "status": "VERIFIED",
-        "phone": "0917499917",
-        "profileImageUrl": "https://d3polnwtp0nqe6.cloudfront.net/default.png",
-        "firstname": "(Apple) Truong",
-        "lastname": "Tran (Employee)",
-        "middlename": "Tien",
-        "gender": "MALE",
-        "role": "COMPANY_EMPLOYEE"
-      },
-      "companyDTO": {
-        "id": "8e407290-9bdb-4e12-b7d3-d1ffdd1d8479",
-        "taxId": "18002752273",
-        "name": "Apple",
-        "address": "Apple",
-        "phone": "0917499917",
-        "email": "email@apple.com",
-        "employeeMaxNum": 5,
-        "websiteUrl": "https://www.apple.com",
-        "status": "REGISTERED",
-        "sizeId": 0,
-        "subCategoryDTOs": [
-          {
-            "id": 57,
-            "name": "NGO/Non-profit",
-            "category": {
-              "id": 12,
-              "name": "Services"
-            }
-          },
-          {
-            "id": 10,
-            "name": "Overseas jobs",
-            "category": {
-              "id": 1,
-              "name": "Type"
-            }
-          },
-          {
-            "id": 14,
-            "name": "Sales",
-            "category": {
-              "id": 2,
-              "name": "Front Office"
-            }
-          }
-        ],
-        "companyBenefitDTOS": [
-          {
-            "id": 51,
-            "description": "$200 a month kindergarten allowance per child under 5 years old",
-            "benefitDTO": {
-              "id": 0,
-              "name": "Kindergarten",
-              "description": "e.g: $200 a month kindergarten allowance per child under 5 years old"
-            },
-            "companyDTO": null
-          },
-          {
-            "id": 52,
-            "description": "Lunch and unlimited snacks",
-            "benefitDTO": {
-              "id": 1,
-              "name": "Canteen",
-              "description": "e.g: Lunch and unlimited snacks"
-            },
-            "companyDTO": null
-          }
-        ],
-        "mediaDTOS": [],
-        "companyLogoURL": "https://d3polnwtp0nqe6.cloudfront.net/default.png",
-        "companyDescription": "Apple đại đế"
-      }
-    }
-  ]
