+++ jobfair-frontend/src/pages/EmployeeRegisterPage/EmployeeRegisterPage.jsx
@@ -1,198 +1,12 @@
-import React, { useState } from "react";
-import { createEmployeesAPI } from "../../services/companyEmployeeService";
-import {
-  Select,
-  Button,
-  Form,
-  Input
-} from 'antd';
-import { tailFormItemLayout, formItemLayout } from "./EmployeeRegisterPage.style";
-import { useSelector } from "react-redux";
-
-const { Option } = Select;
+import EmployeeForm from "../../components/EmployeeForm/EmployeeForm.component";
 
 const RegisterPage = () => {
-  const [form] = Form.useForm();
-  const companyId = useSelector((state) => state.authentication.user.companyId);
-
-
-  const onFinish = (values) => {
-    console.log('Received values of form: ', values);
-
-    createEmployeesAPI({
-      companyId: companyId,
-      confirmPassword: values.confirm,
-      email: values.email,
-      firstName: values.firstName,
-      gender: values.gender,
-      lastName: values.lastName,
-      middleName: values.middleName,
-      password: values.password,
-      phone: values.phone
-    })
-      .then((data) => {
-        console.log(data);
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-  };;
 
   return (
-    <Form
-      {...formItemLayout}
-      form={form}
-      name="register"
-      onFinish={onFinish}
-      scrollToFirstError
-    >
-      <Form.Item
-        name="email"
-        label="E-mail"
-        rules={[
-          {
-            type: 'email',
-            message: 'The input is not valid E-mail!',
-          },
-          {
-            required: true,
-            message: 'Please input your E-mail!',
-          },
-        ]}
-      >
-        <Input />
-      </Form.Item>
-
-      <Form.Item
-        name="password"
-        label="Password"
-        rules={[
-          {
-            required: true,
-            message: 'Please input your password!',
-          },
-        ]}
-        hasFeedback
-      >
-        <Input.Password />
-      </Form.Item>
-
-      <Form.Item
-        name="confirm"
-        label="Confirm Password"
-        dependencies={['password']}
-        hasFeedback
-        rules={[
-          {
-            required: true,
-            message: 'Please confirm your password!',
-          },
-          ({ getFieldValue }) => ({
-            validator(_, value) {
-              if (!value || getFieldValue('password') === value) {
-                return Promise.resolve();
-              }
-
-              return Promise.reject(new Error('The two passwords that you entered do not match!'));
-            },
-          }),
-        ]}
-      >
-        <Input.Password />
-      </Form.Item>
-
-      <Form.Item
-        name="firstName"
-        label="First name"
-        tooltip="What do you want others to call you?"
-        rules={[
-          {
-            required: true,
-            message: 'Please input your nickname!',
-            whitespace: true,
-          },
-        ]}
-      >
-        <Input />
-      </Form.Item>
-
-      <Form.Item
-        name="middleName"
-        label="Middle name"
-        tooltip="What do you want others to call you?"
-        rules={[
-          {
-            required: true,
-            message: 'Please input your nickname!',
-            whitespace: true,
-          },
-        ]}
-      >
-        <Input />
-      </Form.Item>
-
-      <Form.Item
-        name="lastName"
-        label="Last name"
-        tooltip="What do you want others to call you?"
-        rules={[
-          {
-            required: true,
-            message: 'Please input your nickname!',
-            whitespace: true,
-          },
-        ]}
-      >
-        <Input />
-      </Form.Item>
-
-      <Form.Item
-        name="phone"
-        label="Phone Number"
-        rules={[
-          {
-            required: true,
-            message: 'Please input your phone number!',
-          },
-        ]}
-      >
-        <Input
-          style={{
-            width: '100%',
-          }}
-        />
-      </Form.Item>
-
-      <Form.Item
-        name="gender"
-        label="Gender"
-        rules={[
-          {
-            required: true,
-            message: 'Please select gender!',
-          },
-        ]}
-      >
-        <Select placeholder="select your gender">
-          <Option value="MALE">Male</Option>
-          <Option value="FEMALE">Female</Option>
-        </Select>
-      </Form.Item>
-
-      <Form.Item {...tailFormItemLayout}>
-        <Button type="primary" htmlType="submit">
-          Register
-        </Button>
-      </Form.Item>
-    </Form>
+    <>
+      <EmployeeForm/>
+    </>
   );
-};
-
-const handleOnSubmit = (values, actions) => {
-
-  console.log("here");
-
-  
 };
 
 export default RegisterPage;
